+++ pkg/obitax/csvtaxdump_write.go
@@ -1,11 +1,1 @@
 package obitax
-
-import (
-	"git.metabarcoding.org/obitools/obitools4/obitools4/pkg/obiitercsv"
-)
-
-func WriteTaxonomyCSV(iterator ITaxon,
-	terminalAction bool, filenames ...string) *obiitercsv.ICSVRecord {
-
-	return nil
-}